requirements.txt
flask
werkzeug
pillow
rembg
gunicorn